changed ImageService

+++ src/main/java/edu/inai/coursework3/services/ImageService.java
@@ -3,6 +3,7 @@ package edu.inai.coursework3.services;
 import edu.inai.coursework3.util.FileStorageImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -14,10 +15,15 @@ import java.io.IOException;
 public class ImageService {
     private final FileStorageImpl fileStorage;
 
+    @Value("courses.default.image")
+    String defaultAvatar;
+
     @SneakyThrows
     public String saveImage(MultipartFile image, String oldImageName){
         String imageStorageName=fileStorage.save(image.getInputStream(),image.getOriginalFilename());
-        deleteImage(oldImageName);
+        if(!oldImageName.equals(defaultAvatar)){
+            deleteImage(oldImageName);
+        }
         return imageStorageName;
     }
